fix: 尝试修复循环依赖问题
