+++ core/core-frontend/src/custom-component/v-query/Tree.vue
@@ -15,9 +15,8 @@ import {
 import { useEmitt } from '@/hooks/web/useEmitt'
 import { cloneDeep, debounce, sortBy } from 'lodash-es'
 import { getFieldTree } from '@/api/dataset'
-import colorFunctions from 'less/lib/less/functions/color.js'
-import colorTree from 'less/lib/less/tree/color.js'
 import { colorStringToHex } from '@/utils/color'
+import { mixColor } from '@/utils/color'
 
 interface SelectConfig {
   selectValue: any
@@ -390,9 +389,7 @@ const tagColor = computed(() => {
     ? customStyle.background
     : colorStringToHex(customStyle.background)
 
-  return colorFunctions
-    .mix(new colorTree('ffffff'), new colorTree(hexColor.substr(1)), { value: 20 })
-    .toRGB()
+  return mixColor('ffffff', hexColor.substr(1), 0.2)
 })
 </script>
